Support the generation of arith.h with Emscription
When built with the Emscripten toolchain, CMAKE_CROSSCOMPILING_EMULATOR is set
to the Node.js executable. However, the Emscripten runtime does not have
access to the local filesystem by default, so arith.h cannot be written out
directly. Instead, write to the process output stream and write the
CMake-captured output to the file arith.h.

Co-authored-by: Jean-Christophe Fillion-Robin <jchris.fillionr@kitware.com>